+++ superset/assets/javascripts/explore/stores/controls_extended.jsx
