+++ superset/assets/javascripts/explore/stores/visTypes_extended.js
